다음 라운드에도 캐릭터들이 상속되도는 기능 구현 성공함

+++ ApplicationSoftwareProjectTeam2/entities/creatures/Boxer.cs
@@ -65,7 +65,7 @@ namespace ApplicationSoftwareProjectTeam2.entities.creatures
         }
 
         public override EntityTypes getEntityType() => EntityTypes.Avengers;
-        public override byte getLivingEntityId() => 20;
+        public override byte getLivingEntityId() => (byte) 20;
 
         public override void tickAlive()
         {
