+++ ApplicationSoftwareProjectTeam2/entities/creatures/Skulls.cs
@@ -66,7 +66,7 @@ namespace ApplicationSoftwareProjectTeam2.entities.creatures
         #region 캐릭터 아이디 기록
         public override byte getLivingEntityId()
         {
-            return 12;
+            return (byte) 12;
         }
         #endregion
         public override void tickAlive()
